% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{extract.RI.regression.result}
\alias{extract.RI.regression.result}
\title{extract for texreg package}
\usage{
extract.RI.regression.result(model, include.rsquared = FALSE,
  include.adjrs = FALSE, include.nobs = TRUE,
  include.fstatistic = FALSE, include.rmse = FALSE, ...)
}
\description{
extract for texreg package
}
