% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wrappers.R
\name{FRTCI}
\alias{FRTCI}
\title{FRTCI}
\usage{
FRTCI(Y, Z, test.stat, B, gamma, grid.gamma, grid.size,
te.vec, return.matrix, verbose, ...)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector (1=Tx, 0=Co)}

\item{test.stat}{Test statistic function to use on the data.}

\item{B}{Number of permutations to take.}

\item{gamma}{How wide of a CI to make around tau-hat for search}

\item{grid.gamma}{Parameter to govern where the grid points are sampled.  Bigger values means more samples towards the estimated tau-hat.}

\item{grid.size}{Number of points in the grid.}

\item{te.vec}{Vector of taus to examine if you want to override generating ones automatically.}

\item{return.matrix}{TRUE means give back the matrix of all the imputed statistics.  FALSE do not.}

\item{verbose}{TRUE means print out progress bar when fitting and other diagnostics.}

\item{...}{Extra arguments passed to the generate.permutations function (and the test.stat function).}
}
\value{
A list with the value of the test statistic on the observed data, the value of the CI-adjusted p-value,
the plug-in p-value, and other information on the test.
}
\description{
Conduct the FRT CI Method on passed data
}
\examples{
\dontrun{
Z <- rep(c(0, 1), 100)
tau <- 4
Y <- ifelse(Z, rnorm(100, tau), rnorm(100, 0))
tst <- FRTCI(Y, Z)
}
}
