% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wrappers.R
\name{FRTplug}
\alias{FRTplug}
\title{FRTplug}
\usage{
FRTplug(Y, Z, test.stat, tau.hat, ...)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector (1=Tx, 0=Co)}

\item{test.stat}{Test statistic function to use on the data.}

\item{tau.hat}{The value of the plug-in treatment effect. Defaults to the sample average treatment effect.}

\item{...}{Extra arguments passed to the generate.permutations function (and the test.stat function).}
}
\value{
A list with the value of the test statistic on the observed data, the value of the CI-adjusted p-value,
the plug-in p-value, and other information on the test.
}
\description{
Return the plug-in p-value for plugging in tau-hat under the permutation method
Just calls FRTCI.test with the specific, single tau of tau-hat.
}
\examples{
Z <- rep(c(0, 1), 100)
tau <- 4
Y <- ifelse(Z, rnorm(100, tau), rnorm(100, 0))
tst <- FRTplug(Y, Z)
}
