% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{get.p.value}
\alias{get.p.value}
\title{get p-value along with uncertainty on p-value}
\usage{
get.p.value(tst)
}
\arguments{
\item{tst}{A FRTCI.test object from fishpidetect}
}
\value{
p-value and range of p-values due to monte carlo error.
}
\description{
Give confidence bounds (from monte carlo simulation error) for the p-values returned by a test
}
