% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{hettx-deprecated}
\alias{hettx-deprecated}
\alias{test.stat.info}
\alias{KS.stat}
\alias{SKS.stat}
\alias{SKS.stat.cov.pool}
\alias{SKS.stat.cov}
\alias{SKS.stat.int.cov.pool}
\alias{SKS.stat.int.cov}
\alias{SKS.stat.cov.rq}
\alias{rq.stat}
\alias{rq.stat.cond.cov}
\alias{rq.stat.uncond.cov}
\alias{WSKS.t}
\alias{SKS.pool.t}
\alias{make.linear.data}
\alias{make.quadradic.data}
\alias{make.skew.data}
\alias{make.randomized.dat}
\alias{make.randomized.compliance.dat}
\alias{get.p.value}
\alias{variance.ratio.test}
\title{Deprecated functions in hettx}
\usage{
test.stat.info()

KS.stat(...)

SKS.stat(...)

SKS.stat.cov.pool(...)

SKS.stat.cov(...)

SKS.stat.int.cov.pool(...)

SKS.stat.int.cov(...)

SKS.stat.cov.rq(...)

rq.stat(...)

rq.stat.cond.cov(...)

rq.stat.uncond.cov(...)

WSKS.t(...)

SKS.pool.t(...)

make.linear.data(...)

make.quadradic.data(...)

make.skew.data(...)

make.randomized.dat(...)

make.randomized.compliance.dat(...)

get.p.value(...)

variance.ratio.test(...)
}
\arguments{
\item{...}{Arguments passed to the replacement function.}
}
\description{
These functions have been renamed to use snake_case. The old dot-case
names still work but will emit a deprecation warning. Please update
your code to use the new names.
}
\keyword{internal}
