% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fake_data.R
\name{make_randomized_compliance_dat}
\alias{make_randomized_compliance_dat}
\title{Generate fake data with noncompliance.}
\usage{
make_randomized_compliance_dat(
  n,
  p = 0.6,
  science.table.generator = make_linear_data,
  include.POs = TRUE,
  ...
)
}
\arguments{
\item{n}{Sample size}

\item{p}{Proportion treated}

\item{science.table.generator}{Method to generate potential outcomes}

\item{include.POs}{Preserve potential outcomes in returned value}

\item{...}{To be passed to science.table.generator}
}
\value{
Data frame with data randomized to tx and control, and compliers, etc.
}
\description{
This will generate and randomize a science table to get observed outcomes and treatment assignment
}
\seealso{
make_randomized_dat
}
