% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fake_data.R
\name{make_randomized_dat}
\alias{make_randomized_dat}
\title{Make fake data for simulations}
\usage{
make_randomized_dat(
  n,
  p = 0.6,
  science.table.generator = make_linear_data,
  include.POs = TRUE,
  as.data.frame = TRUE,
  ...
)
}
\arguments{
\item{n}{Sample size}

\item{p}{Proportion treated}

\item{science.table.generator}{Data generator}

\item{include.POs}{TRUE/FALSE. Keep POs}

\item{as.data.frame}{TRUE/FALSE.  Return as dataframe or as list of elements.}

\item{...}{Additional to be passed to science.table.generator}
}
\value{
Either a list of elements or a dataframe.
}
\description{
Randomize a science table to get observed outcomes and treatment assignment
}
