% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Penn46_ascii}
\alias{Penn46_ascii}
\title{Sample data set}
\format{A dataframe containing 6384 observations and 12 columns.}
\usage{
Penn46_ascii
}
\description{
This is a sample data set to illustrate the package methods.
}
\keyword{dataset}
