% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{print.FRTCI.test}
\alias{print.FRTCI.test}
\title{print.FRTCI.test}
\usage{
\method{print}{FRTCI.test}(x, ...)
}
\arguments{
\item{x}{An object of class \code{FRTCI.test}}

\item{...}{Further arguments to be passed to \code{print.FRTCI.test()}}
}
\description{
Print out FRTCI.test object.
}
