% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_estimation_library.R
\name{print.RI.R2.result}
\alias{print.RI.R2.result}
\title{Print out results on estimating treatment effect R2}
\usage{
\method{print}{RI.R2.result}(x, ...)
}
\arguments{
\item{x}{Object to print}

\item{...}{Ignored}
}
\description{
Print various R2 measures including the bounds due to the sensitivity analysis.
}
