% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_stat_calc.R
\name{rq.stat}
\alias{rq.stat}
\title{rq.stat}
\usage{
rq.stat(Y, Z, rq.pts)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector}

\item{rq.pts}{Sequence of quantile points at which to evaluate the test. Default is seq(.1, .9, by = .1). Should not go beyond 0 and 1.}
}
\value{
The value of the test.
}
\description{
Kolmogorov-smirnov statistic via quantile regression with covariates
}
