% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_calc_helpers.R
\name{rq.stat.uncond.cov}
\alias{rq.stat.uncond.cov}
\title{rq.stat.uncond.cov}
\usage{
rq.stat.uncond.cov(Y, Z, X, rq.pts)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector}

\item{X}{Additional pre-treatment covariates to adjust for in estimation, but not to interact with treatment.}

\item{rq.pts}{Sequence of quantile points at which to evaluate the test. Default is seq(.1, .9, by = .1). Should not go beyond 0 and 1.}
}
\value{
The value of the test.
}
\description{
Kolmogorov-smirnov statistic via quantile regression with covariates
Unconditional approach; see Firpo (2007)
}
