% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_systematic.R
\name{SE}
\alias{SE}
\title{Extract the standard errors from a var-cov matrix.}
\usage{
SE(object, ...)
}
\arguments{
\item{object}{est.beta object}

\item{...}{unused}
}
\description{
Extract the standard errors from a var-cov matrix.
}
\examples{
df <- make.randomized.dat( 1000, gamma.vec=c(1,1,1,2), beta.vec=c(-1,-1,1,0) )
es <- estimate_systematic( Yobs ~ Z,  interaction.formula = ~ A + B, data = df )
SE(es)

}
