% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_stat_calc.R
\name{SKS.stat.cov.rq}
\alias{SKS.stat.cov.rq}
\title{SKS.stat.cov.rq}
\usage{
SKS.stat.cov.rq(Y, Z, X)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector}

\item{X}{Additional pre-treatment covariates to adjust for in estimation, but not to interact with treatment. Default is NULL.}
}
\value{
The value of the test.
}
\description{
Shifted kolmogorov-smirnov statistic with covariates and rq.
}
