% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_stat_calc.R
\name{SKS.stat}
\alias{SKS.stat}
\title{SKS.stat}
\usage{
SKS.stat(Y, Z)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector}
}
\value{
The value of the test.
}
\description{
Shifted kolmogorov-smirnov statistic. Calculate KS distance between Y0 and Y1
shifted by sample tau.
}
\examples{
df <- make.randomized.dat( 1000, gamma.vec=c(1,1,1,2), beta.vec=c(-1,-1,1,0) )
SKS.stat(df$Yobs, df$Z)

}
\seealso{
KS.stat, SKS.stat.cov

detect_idiosyncratic
}
