% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_stat_calc.R
\name{SKS_stat}
\alias{SKS_stat}
\title{SKS_stat}
\usage{
SKS_stat(Y, Z)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector}
}
\value{
The value of the test.
}
\description{
Shifted kolmogorov-smirnov statistic. Calculate KS distance between Y0 and Y1
shifted by sample tau.
}
\examples{
df <- make_randomized_dat( 1000, gamma.vec=c(1,1,1,2), beta.vec=c(-1,-1,1,0) )
SKS_stat(df$Yobs, df$Z)

}
\seealso{
KS_stat, SKS_stat_cov

detect_idiosyncratic
}
