% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_stat_calc.R
\name{SKS_stat_cov_rq}
\alias{SKS_stat_cov_rq}
\title{SKS_stat_cov_rq}
\usage{
SKS_stat_cov_rq(Y, Z, X)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector}

\item{X}{Additional pre-treatment covariates to adjust for in estimation, but not to interact with treatment.}
}
\value{
The test statistic value.
}
\description{
Shifted kolmogorov-smirnov statistic with covariates and quantile regression.
}
\examples{
df <- make_randomized_dat( 1000, gamma.vec=c(1,1,1,2), beta.vec=c(-1,-1,1,0) )
SKS_stat_cov_rq(df$Yobs, df$Z, df$A)

}
