% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_stat_calc.R
\name{test.stat.info}
\alias{test.stat.info}
\title{test.stat.info}
\usage{
test.stat.info()
}
\description{
A list of test statistics for detect.idiosyncratic(), and information on use cases when each is appropriate.
}
\examples{
test.stat.info()

}
