% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_stat_calc.R
\name{test_stat_info}
\alias{test_stat_info}
\title{test_stat_info}
\usage{
test_stat_info()
}
\description{
A list of test statistics for detect.idiosyncratic(), and information on use cases when each is appropriate.
}
\examples{
test_stat_info()

}
