% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_tidy.R
\name{tidy.RI.R2.result}
\alias{tidy.RI.R2.result}
\title{Tidy an RI.R2.result}
\usage{
\method{tidy}{RI.R2.result}(x, ...)
}
\arguments{
\item{x}{An \code{RI.R2.result} object from \code{R2()}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A data.frame with R-squared bound estimates. For ITT results,
  one row. For LATE results, three rows (Compliers, Noncompliers,
  Covariates and compliers).
}
\description{
Tidy an RI.R2.result
}
