% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_systematic.R
\name{variance_ratio_test}
\alias{variance_ratio_test}
\title{Variance ratio test}
\usage{
variance_ratio_test(Yobs, Z, data = NULL)
}
\arguments{
\item{Yobs}{Outcome}

\item{Z}{Treatment assignment vector}

\item{data}{Dataframe with variables listed in formula and control.formula}
}
\description{
Given vector of observed outcomes and treatment vector, test to see if there is evidence
the variances are different (taking kurtosis into account).
}
\examples{
df <- make_randomized_dat( 1000, gamma.vec=c(1,1,1,2), beta.vec=c(-1,-1,1,0) )
variance_ratio_test(df$Yobs, df$Z)

}
