% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_systematic.R
\name{vcov.RI.regression.result}
\alias{vcov.RI.regression.result}
\title{Get vcov() from object.}
\usage{
\method{vcov}{RI.regression.result}(object, ...)
}
\arguments{
\item{object}{est.beta object}

\item{...}{unused}
}
\description{
Get vcov() from object.
}
\examples{
df <- make.randomized.dat( 1000, gamma.vec=c(1,1,1,2), beta.vec=c(-1,-1,1,0) )
es <- estimate_systematic( Yobs ~ Z,  interaction.formula = ~ A + B, data = df )
vcov(es)

}
