% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_systematic.R
\name{coef.RI.regression.result}
\alias{coef.RI.regression.result}
\title{Extract coefficients of a fit RI regression model.}
\usage{
\method{coef}{RI.regression.result}(object, ...)
}
\arguments{
\item{object}{A RI.regression.result object.}

\item{...}{Unused}
}
\description{
Extract coefficients of a fit RI regression model.
}
