% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fake_data.R
\name{make.linear.data}
\alias{make.linear.data}
\alias{make.quadradic.data}
\alias{make.skew.data}
\title{Generate dataset according to a linear model.}
\usage{
make.linear.data(n, gamma.vec = c(1, 2, 2, 1), gamma2.vec = NULL,
  beta.vec = c(-1, -1, 1), ideo.sd = 0, quad.tx = FALSE,
  stand.norm.X = FALSE, corr.X = !stand.norm.X)

make.quadradic.data(n, beta.vec = c(-1, -1, 1))

make.skew.data(n, beta.vec = c(-1, -1, 1))
}
\arguments{
\item{n}{Sample size}

\item{gamma.vec}{Control outcome surface}

\item{gamma2.vec}{Quadratic terms}

\item{beta.vec}{Treatment effect surface}

\item{ideo.sd}{Ideosyncratic residual variation}

\item{quad.tx}{Quadratic treatment effects?}

\item{stand.norm.X}{Xs shifted around 0 or no}

\item{corr.X}{TRUE or FALSE.  Have Xs correlated or no.}
}
\value{
List of elements of data (not data frame)
}
\description{
Given the parameters, generate a dataset and return a potential outcomes schedule (science table)
of synthetic potential outcomes.
}
\section{Functions}{
\itemize{
\item \code{make.quadradic.data}: Generate dataset according to a quadratic model

\item \code{make.skew.data}: Generate dataset with a skew
}}

