% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{plot.FRTCI.test}
\alias{plot.FRTCI.test}
\title{plot.FRTCI.test}
\usage{
\method{plot}{FRTCI.test}(x, true.tau, xlab, ylab, true.tau.col, plot.envelope, ci.line.col, ...)
}
\arguments{
\item{x}{An object of class \code{FRTCI.test}}

\item{true.tau}{The true value of tau, if known. Default is NULL.}

\item{xlab}{X-axis label. Default is tau.}

\item{ylab}{Y-axis label. Default is "p-value".}

\item{true.tau.col}{Color to plot true tau value, if provided. Default is red.}

\item{plot.envelope}{Plot envelope around tested values of tau. Default is TRUE.}

\item{ci.line.col}{Color to plot confidence interval around estimated treatment effect. Default is blue.}

\item{...}{Further arguments to be passed to \code{print.FRTCI.test()}}
}
\description{
Plot curve from FRTCI.test object.
}
