% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_stat_calc.R
\name{SKS.stat.int.cov}
\alias{SKS.stat.int.cov}
\title{SKS.stat.int.cov}
\usage{
SKS.stat.int.cov(Y, Z, W, X)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector}

\item{W}{Additional pre-treatment covariates to interact with T to define linear model of treatment effects.}

\item{X}{Additional pre-treatment covariates to adjust for in estimation, but not to interact with treatment. Default is NULL.}
}
\value{
The value of the test.
}
\description{
Shifted kolmogorov-smirnov statistic with a linear treatment
effect model defined by W
}
\details{
This will attempt to remove any systematic variation corresponding
to W and then return a SKS statistic on the residuals to measure
any variation "left over".

X are _additional_ covariates to adjust for beyond those involved
in treatment effect model.  It will automatically ajust for W as
well.  Do not put a covariate in for both X and W.

This is the test statistic used in Ding, Feller, and Miratrix (2016), JRSS-B.

This method first adjusts for baseline and then models treatment effect
on the residuals to not split treatment effects.

We recommend this method over the "pool" method.
}
