% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_stat_calc.R
\name{SKS.stat}
\alias{SKS.stat}
\title{SKS.stat}
\usage{
SKS.stat(Y, Z)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector}
}
\value{
The value of the test.
}
\description{
Shifted kolmogorov-smirnov statistic. Calculate KS distance between Y0 and Y1
shifted by sample tau.
}
