% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_systematic.R
\name{vcov.RI.regression.result}
\alias{vcov.RI.regression.result}
\title{Get vcov() from object.}
\usage{
\method{vcov}{RI.regression.result}(object, ...)
}
\arguments{
\item{object}{est.beta object}

\item{...}{unused}
}
\description{
Get vcov() from object.
}
