% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_stat_calc.R
\name{WSKS.t}
\alias{WSKS.t}
\title{WSKS.t}
\usage{
WSKS.t(Y, Z, W)
}
\arguments{
\item{Y}{Observed outcome vector}

\item{Z}{Treatment assigment vector}

\item{W}{A a factor or categorical covariate.}
}
\value{
The value of the test.
}
\description{
Weighted average of the group-level SKS statistics
}
